Move backgroundStyle to generalContent store
