+++ src/store/index.ts
@@ -1,6 +1,10 @@
 import { currentUser, initCurrentUser, updateCurrentUser } from './currentUser';
-import { generalContent, initGeneralContent } from './generalContent';
+import {
+  generalContent,
+  initGeneralContent,
+  backgroundStyle,
+} from './generalContent';
 
 export const initStore = Promise.all([initCurrentUser, initGeneralContent]);
 
-export { currentUser, updateCurrentUser, generalContent };
+export { currentUser, updateCurrentUser, generalContent, backgroundStyle };
